+++ yt_fts/search.py
@@ -1,6 +1,5 @@
 from pprint import pprint
-from .download import get_channel_id_from_input
-from .db_utils import * 
+from .db_utils import *
 from .utils import *
 from rich.console import Console
 from rich.text import Text
